pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Niccolò Cantù", email = "niccolo.cantu@nina.no"},
]
dependencies = [
    "titiler-application>=0.17.0",
    "uvicorn>=0.27.0.post1",
    "matplotlib>=3.8.2",
    "titiler-extensions[cogeo]>=0.17.0",
    "rio-tiler>=6.4.0",
    "scipy>=1.12.0",
    "ipdb>=0.13.13",
    "opencv-python>=4.9.0.80",
    "jupyter>=1.0.0",
    "notebook>=7.0.7",
    "jupyterlab>=4.1.0",
    "walrus>=0.9.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

